Sets up and runs GSEA using population-based FST values

Sets up and runs GSEA using population-based FST values

setupGSEArun(
  fst_file,
  annotation_file,
  snp2gene_file,
  SET_PERM = 10000,
  SNP2GENE_DIST = 5e+05,
  MIN_GENE = 10,
  MAX_GENE = 300,
  SET_SEED = 42,
  output_folder
)

Arguments

fst_file
(char) path to file with SNP-level FST statistics.

annotation_file
(char) path to pathway definitions GMT file.

snp2gene_file
(char) path to file with snp2gene mappings (output of SNP2gene.R).

SET_PERM
(integer) set cycle of permutations to run (default=10000).

SNP2GENE_DIST
(integer) value for GSEA --distance. Max. distance between SNP and gene for their association in snp2gene_file (default=500000).

MIN_GENE
(integer) value for GSEA --setmin. Min. number of genes in a gene set to be considered (default=10).

MAX_GENE
(integer) value for GSEA --setmax. Max. number of genes in a gene set to be considered (default=300).

SET_SEED
(integer) value for GSEA --seed.

output_folder
(char) path to store output files.

Value

none